+++ tv_project/src/App.js
@@ -1,3 +1,4 @@
+import {useState, useEffect} from 'react';
 import { BrowserRouter as Router, Routes, Route} from 'react-router-dom'
 import './App.css';
 
@@ -8,10 +9,25 @@ import Show from './pages/Show';
 
 
 function App() {
+  const [shows, setShows] = useState([]);
+
+	const fetchShows = () => {
+		fetch("https://api.tvmaze.com/shows")
+			.then((response) => {
+				return response.json();
+			})
+			.then((data) => {
+				setShows(data);
+			});
+	};
+
+	useEffect(() => {
+		fetchShows();	
+	},[]);
   return (  
       <Router>
         <Routes>
-          <Route path='/' element={<Home/>}/>
+          <Route path='/' element={<Home shows={shows}/>}/>
           <Route path='/about' element={<About/>}/>
           <Route path='/contact' element={<Contact/>}/>
           <Route path='/show' element={<Show/>}/>
