+++ tv_project/src/components/Cards.css
@@ -1,0 +1,4 @@
+.shows{
+    padding: 0 20px;
+    height:fit-content;
+}
