+++ tv_project/src/components/SearchBar.css
@@ -1,4 +1,5 @@
 .searchbar{
+    margin-top: 70px;
     height:30px;
     background-color:lightgray;
     display: flex;
