+++ tv_project/src/pages/About.css
@@ -1,3 +1,6 @@
-.about{
+
+h4{
+    margin-top: 70px;
+    color:black;
     padding: 10px 20px;
 }
