+++ tv_project/src/pages/Contact.css
@@ -1,3 +1,4 @@
-.contact{
-    padding: 10px 20px;
+h4 {
+	margin-top: 70px;
+	color: black;
 }
