+++ tv_project/src/pages/Home.css
@@ -1,4 +1,4 @@
-
-h4{
-    padding: 10px 20px;
+.home h4{
+   margin:0;
+    color:black;
 }
