+++ tv_project/src/pages/Show.css
@@ -1,3 +1,5 @@
-.show{
+h4{
+    margin-top: 70px;
+    color:black;
     padding: 10px 20px;
 }
